+++ Graficas - MM8BExamen/nbproject/project.properties
@@ -54,7 +54,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=graficas.mm8bexamen.GraficasMM8BExamen
+main.class=graficas.GraficasMM8BExamen
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
