+++ Graficas - MM8BExamen/src/graficas/GraficasMM8BExamen.java
@@ -1,0 +1,23 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package graficas;
+
+/**
+ *
+ * @author demg
+ */
+public class GraficasMM8BExamen {
+
+    /**
+     * @param args the command line arguments
+     */
+    public static void main(String[] args) {
+        // TODO code application logic here
+        Main m = new Main();
+        m.setVisible(true);
+    }
+    
+}
